run-ci.sh: disable running in older releases
An issue arose recently where the run-ci.sh script within entity-manager
started identifying json formatting issues due to the docker container
grabbing a newer version of a package.

With our older 1020, 1030, 1040 releases the goal is to do as much as we
can without hacking things too much. Given the limited use of the
run-ci.sh add-on scripts and issues like this, disable running of these
scripts in older releases.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>