build-unit-test-docker: fix clang version argument
The latest version of bash seems to not pass additional arguments
when called with `-c` unless those arguments are after the `--`.

This caused the build-unit-test-docker to not install the latest
version of clang, but instead the release of clang that was present
in the Ubuntu repository.

Add "19" after `--` so that the latest versions of clang tools are
installed.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I894b77b2a03e9125df50a297c9898b3456846cdc